refactor: canonicalize outbound rate limiters
